qa/qa_pytest/requirements.txt
allure_python_commons==2.9.45
arrow==1.2.2
beautifulsoup4==4.9.3
fastapi==0.115.6
openpyxl==3.0.7
pandas==1.2.4
pluggy==0.13.1
psutil==5.8.0
pydantic==2.10.3
pytest==6.2.3
Requests==2.32.3
ruamel.base==1.0.0
six==1.15.0
urllib3==1.26.4
uvicorn==0.34.0
xlrd==2.0.1
xlrd==1.2.0
